fix(dotnet): DesktopHost 里 nint | long 的运算结果是 long，补回显式转换
CS1503 出在 MarkToolWindow：`nint | long` 按 C# 规则提升为 long，
而 SetWindowLongPtr 第 3 参是 nint，隐式转换不存在。加回 (nint) 并把原因写在注释里。

+++ dotnet/Tjt.App/Win32/DesktopHost.cs
@@ -119,7 +119,8 @@ internal static class DesktopHost
     public static void MarkToolWindow(nint hwnd)
     {
         var style = NativeMethods.GetWindowLongPtr(hwnd, NativeMethods.Constants.GwlExStyle);
-        var updated = style | NativeMethods.Constants.WsExToolWindow;
+        // nint | long 的运算结果是 long，而 SetWindowLongPtr 第 3 参是 nint —— 必须显式转回来
+        var updated = (nint)(style | NativeMethods.Constants.WsExToolWindow);
         if (updated != style)
         {
             NativeMethods.SetWindowLongPtr(hwnd, NativeMethods.Constants.GwlExStyle, updated);
